Update scss and property renames
