--- src/constants/constants.scss
@@ -1,93 +1,0 @@
-
-/* V A R I A B L E S */
-// colors
-$color__gray--lightest: #f3f3f3 !default;
-$color__gray--light: #eeeeee !default;
-$color__gray--border: #dddddd !default;
-$color__gray--medium: #858585 !default;
-$color__gray--dark: #555555 !default;
-$color__blue: #4ca1ff !default;
-$color__blue--dark: #003c71 !default;
-$color__green: #4cd964 !default;
-$color__yellow: #ffbd3b !default;
-$color__red: #ff5556 !default;
-$color__red--dark: #ed2e2f !default;
-$color__white: #ffffff !default;
-$color__black: #171719 !default;
-$color__gradient--blue: linear-gradient(180deg, #6d8dff 0%, #4cb9ef 100%)
-  !default;
-$color__gradient--nav: linear-gradient(180deg, #4cb9ef 0%, #6d8dff 100%)
-  !default;
-$color__gradient--yellow: linear-gradient(180deg, #ffaa55 0%, #ffbd3b 100%)
-  !default;
-$color__gradient--red: linear-gradient(to top, #ff7c5b 0%, #ed4c4c 100%)
-  !default;
-// text/typography
-// measurements
-$border-radius: 8px;
-$border-radius--btn: 20px;
-$border-radius--card: 10px;
-$ticket__cutout-border--size: 4px;
-$ticket__cutout-border--radius: 8px;
-$ticket__cutout-size--height: 18px;
-$ticket__cutout-size--width: 6px;
-$ticket__cutout-sides: 23px;
-$width__nav: 15em;
-$width__wrapper: 800px;
-// media queries
-$mq--desktop: 1200px;
-/* end variables */
-
-/* F U N C T I O N S */
-@function emify($target, $context) {
-  @return ($target/$context) * 1em;
-}
-/* end functions */
-
-/* M I X I N S */
-@mixin clearfix {
-  &:after {
-    content: "";
-    display: table;
-    height: 0;
-    clear: both;
-    visibility: hidden;
-  }
-}
-
-@mixin dialog_shadow {
-  box-shadow: 0 2px 3px rgba(0, 0, 0, 0.1);
-}
-
-@mixin ticket_cutout($position) {
-  #{$position}: -$ticket__cutout-size--width;
-  &::after {
-    background: rgba(0, 0, 0, 0);
-    border: 5px solid $color__white;
-    border-radius: 50%;
-    content: "";
-    height: $ticket__cutout-sides;
-    position: absolute;
-    #{$position}: -12px;
-    top: -2px;
-    width: $ticket__cutout-sides;
-  }
-}
-
-@mixin flex_align($position) {
-  align-items: center;
-  display: flex;
-  justify-content: $position;
-}
-
-@mixin hidden_scrolling {
-  box-sizing: content-box;
-  overflow-y: auto;
-  // overflow: -moz-scrollbars-none;
-  -ms-overflow-style: none;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
-/* end mixins */
