+++ src/scss/variables.scss
@@ -4,14 +4,19 @@
 // colors
 $brandBlue: #053d6f;
 $gray: #8e8e93;
+$mediumGray: #858585 !default;
 $blue: #51a3fc;
+$lighterBlue: #4ca1ff !default;
+$darkBlue: #003c71 !default;
 $orange: #fdbc49;
 $red: #fc575a;
 $darkGray: #555555;
 $lightGray: #d1d1d1;
 $lighterGray: #eeeeee;
 $white: #FFF;
+$black: black;
 $errorRed: red;
+
 $breakpointMobile: 1200px;
 
 $gradient: linear-gradient(to bottom, #6f90fc 0%, #52baed 100%);
